# Conflicts:
#	src/MarginTrading.Common/BackendContracts/OrderContract.cs
#	src/MarginTrading.Common/BackendContracts/OrderFullContract.cs
#	src/MarginTrading.Core/IClientNotifyService.cs
#	src/MarginTrading.Core/IRabbitMqNotifyService.cs
#	src/MarginTrading.Services/RabbitMqNotifyService.cs

+++ src/MarginTrading.Common/BackendContracts/OrderFullContract.cs
@@ -7,16 +7,16 @@ namespace MarginTrading.Common.BackendContracts
     public class OrderFullContract : OrderContract
     {
         public string TradingConditionId { get; set; }
-        public double QuoteRate { get; set; }
+        public decimal QuoteRate { get; set; }
         public int AssetAccuracy { get; set; }
-        public double CommissionLot { get; set; }
+        public decimal CommissionLot { get; set; }
         public DateTime? StartClosingDate { get; set; }
         public OrderFillType FillType { get; set; }
         public string Comment { get; set; }
-        public double InterestRateSwap { get; set; }
-        public double MarginInit { get; set; }
-        public double MarginMaintenance { get; set; }
-        public double OpenCrossPrice { get; set; }
-        public double CloseCrossPrice { get; set; }
+        public decimal InterestRateSwap { get; set; }
+        public decimal MarginInit { get; set; }
+        public decimal MarginMaintenance { get; set; }
+        public decimal OpenCrossPrice { get; set; }
+        public decimal CloseCrossPrice { get; set; }
     }
 }
