+++ src/MarginTrading.Services/RabbitMqNotifyService.cs
@@ -27,7 +27,7 @@ namespace MarginTrading.Services
             _publishers = publishers;
             _log = log;
         }
-        public Task AccountHistory(string accountId, string clientId, double amount, double balance, double withdrawTransferLimit, AccountHistoryType type, string comment = null)
+        public Task AccountHistory(string accountId, string clientId, decimal amount, decimal balance, decimal withdrawTransferLimit, AccountHistoryType type, string comment = null)
         {
             var record = new MarginTradingAccountHistory
             {
@@ -99,7 +99,7 @@ namespace MarginTrading.Services
             return TryProduceMessageAsync(_settings.RabbitMqQueues.AccountChanged.ExchangeName, message);
         }
 
-        public Task AccountStopout(string clientId, string accountId, int positionsCount, double totalPnl)
+        public Task AccountStopout(string clientId, string accountId, int positionsCount, decimal totalPnl)
         {
             var message = new { clientId, accountId, positionsCount, totalPnl };
             return TryProduceMessageAsync(_settings.RabbitMqQueues.AccountStopout.ExchangeName, message);
